section-2-tracking-a-dl-pipeline-at-scale/chapter-4-tracking-code-and-data-versioning/requirements.txt
boto3
ipykernel
ipywidgets
mlflow
pandas
python-dotenv
scikit-learn
torch
torchaudio
torchinfo
torchmetrics
torchvision
tqdm
transformers
click